Corrige bug ao tentar ativar o primeiro alerta

+++ components/ChatBot.js
@@ -94,12 +94,12 @@ export async function ChatBot(message) {
                 // Ativa ou desativa o alerta
                 if (action) {
                     // Se o alerta já estiver ativado, não faz nada
-                    if (user.alerts.some(alert => alert.AlertId === alerts[i].id)) return 'O alerta para ```' + name + '``` já está ativado!'
+                    if (user.alerts !== undefined && user.alerts.some(alert => alert.AlertId === alerts[i].id)) return 'O alerta para ```' + name + '``` já está ativado!'
                     // Se o alerta não estiver ativo, ativa
                     await AlertUser.create({UserId: message.from, AlertId: alerts[i].id})
                 } else if (!action) {
                     // Se o alerta já estiver desativado, não faz nada
-                    if (!user.alerts.some(alert => alert.AlertId === alerts[i].id)) return 'O alerta para ```' + name + '``` já está desativado!'
+                    if (user.alerts === undefined || !user.alerts.some(alert => alert.AlertId === alerts[i].id)) return 'O alerta para ```' + name + '``` já está desativado!'
                     // Se o alerta estiver ativo, desativa
                     await AlertUser.destroy({where: {UserId: message.from, AlertId: alerts[i].id}})
                 }
